Supercharging superbooga: 2
